Speakers rework (first part) Speaker now is a pamello entity, so it has id and repository But still no events so remote entities can't be updated, anyway it's unnecessary for now because even bigger audio rework is coming

+++ PamelloV7.Server/Config/Root/Discord/Commands/Commands.cs
@@ -3,7 +3,7 @@ namespace PamelloV7.Server.Config.Root.Discord.Commands;
 public class Commands : IConfigNode
 {
     public bool GlobalRegistration { get; set; } = false;
-    public int[] GuildsIds { get; set; } = [];
+    public ulong[] GuildsIds { get; set; } = [];
     public Aliases.Alias[] Aliases { get; set; } = [];
     
     public void EnsureRight() {
